Implemented AVC-116: Parent app: Details screens for other feed types

+++ source/Avecare.code/Common/Networking/SyncEngine/syncOrganizationTemplates.swift
@@ -24,35 +24,36 @@ extension SyncEngine {
         notifySyncStateChanged(message: "Syncing down 🔻 organization templates")
 
         // Sync down from server and update our local DB.
-        if appSession.userProfile.isSupervisor,
-           let unitId = RLMSupervisor.details?.primaryUnitId,
-            let unitDetails = RLMUnit.details(for: unitId),
-            let institutionDetails = RLMInstitution.details(for: unitDetails.institutionId) {
-            OrganizationsAPIService.getOrganizationLogTemplates(id: institutionDetails.organizationId) { [weak self] result in
-                switch result {
-                case .success(let templates):
-                    // TODO: compare versions, and id's, to preserve old templates?
-                    // for now, delete all existing templates
-                    RLMFormTemplate.findAll().forEach({
-                        $0.clean()
-                        $0.delete()
-                    })
+        if appSession.userProfile.isSupervisor {
+            if let unitId = RLMSupervisor.details?.primaryUnitId,
+                let unitDetails = RLMUnit.details(for: unitId),
+                let institutionDetails = RLMInstitution.details(for: unitDetails.institutionId) {
+                OrganizationsAPIService.getOrganizationLogTemplates(id: institutionDetails.organizationId) { [weak self] result in
+                    switch result {
+                    case .success(let templates):
+                        // TODO: compare versions, and id's, to preserve old templates?
+                        // for now, delete all existing templates
+                        RLMFormTemplate.findAll().forEach({
+                            $0.clean()
+                            $0.delete()
+                        })
 
-/*                  // Will we ever have more then one org. at a time?
-                    guard let organization = RLMOrganization.details(for: institutionDetails.organizationId) else {
-                        fatalError()
-                    }
-                    // link with organization  (inverse relationship).  Is this even needed???
-                    templates.forEach({ $0.organization = organization })
+/*                        // Will we ever have more then one org. at a time?
+                        guard let organization = RLMOrganization.details(for: institutionDetails.organizationId) else {
+                            fatalError()
+                        }
+                        // link with organization  (inverse relationship).  Is this even needed???
+                        templates.forEach({ $0.organization = organization })
 */
-                    // save downloaded template(s)
-                    RLMFormTemplate.createOrUpdateAll(with: templates)
-                    DDLogDebug("⬇️ DOWN syncComplete!  Total \'\(RLMFormTemplate.className())\' items in DB: \(RLMFormTemplate.findAll().count)")
-                    self?.syncStates[syncKey] = .complete
-                    syncCompletion(nil)
-                case .failure(let error):
-                    self?.syncStates[syncKey] = .complete
-                    syncCompletion(error)
+                        // save downloaded template(s)
+                        RLMFormTemplate.createOrUpdateAll(with: templates)
+                        DDLogDebug("⬇️ DOWN syncComplete!  Total \'\(RLMFormTemplate.className())\' items in DB: \(RLMFormTemplate.findAll().count)")
+                        self?.syncStates[syncKey] = .complete
+                        syncCompletion(nil)
+                    case .failure(let error):
+                        self?.syncStates[syncKey] = .complete
+                        syncCompletion(error)
+                    }
                 }
             }
         } else {
